The BufferBuilder passed to the entity renderers seems to expect quads; draws everything at once after all entities have pushed to the buffer. Will probably have to intercept the world render event and use a custom render setup.

+++ src/main/java/jigglyslimes/JigglySlimes.java
@@ -1,34 +1,28 @@
 package jigglyslimes;
 
-import net.minecraft.block.Block;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.monster.SlimeEntity;
 import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.event.RegistryEvent;
 import net.minecraftforge.event.entity.living.LivingDeathEvent;
 import net.minecraftforge.event.entity.living.LivingEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.client.registry.RenderingRegistry;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
-import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
-import net.minecraftforge.fml.event.lifecycle.InterModProcessEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.util.WeakHashMap;
-import java.util.stream.Collectors;
 
 @Mod(JigglySlimes.MODID)
 public class JigglySlimes {
 
     public static final String MODID = "jigglyslimes";
-    public static final double GRAVITY = -32.0; // In m/s^2
-    public static final double AIR_FRICTION = 50.0;
-    public static final double AIR_DENSITY = 1.2; // In kg/m^3
+    public static final float GRAVITY = -32.0F; // In m/s^2
+    public static final float AIR_FRICTION = 50.0F;
+    public static final float AIR_DENSITY = 1.2F; // In kg/m^3
 
     private static final Logger LOGGER = LogManager.getLogger();
